backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pillow>=10.0
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
django-storages>=1.14
dj-database-url>=2.1
